requirements.txt
requests
google-generativeai
boto3
moviepy